refactor(types): move twitter types in their own file
